LargestProductinaSeriesLargeseries.cs fix

+++ project_euler/LargestProductinaSeriesLargeseries.cs
@@ -2,13 +2,13 @@
 
 public class LargestProductinaSeriesLargeseries
 {
-    public static int LargestProductinaSeries(string series, int adjacent)
+    public static long LargestProductinaSeries(string series, int adjacent)
     {
-        int largestProduct = 0;
+        long largestProduct = 0;
         int length = series.Length;
         for (int i = 0; i < length-adjacent; i++)
         {
-            int tempProduct = 1;
+            long tempProduct = 1;
             for (int j = 0; j < adjacent; j++)
             {
                 tempProduct *= series[i+j] - '0';
